add library folder details screen This commit includes a major refactoring of the library, abstracting the core view model and UI events so they can be reused for both the main and details screen.
