+++ app/src/main/java/app/musikus/recorder/presentation/RecorderUi.kt
@@ -750,7 +750,7 @@ private fun PreviewRecorderUi(
                 ),
                 currentPlaybackRawMedia = null
             ),
-            eventHandler = {},
+            eventHandler = { true },
             mediaController = null,
             playerState = null,
             currentPlaybackPosition = 0,
